# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API en la cual podemos consultar cotizaciones de diferentes monedas",
        "title": "Criptomoneda API",
        "contact": {}
    },
    "paths": {
...
        "/usuarios": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Lista de usuario registrados",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Usuario"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nombre de usuario elegido",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "nombre del nuevo usuario",
                        "name": "nombre",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "apellido del nuevo usuario",
                        "name": "apellido",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email del nuevo usuario",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dni | cedula | pasaporte",
                        "name": "tipo_documento",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "identificador de documento elegido",
                        "name": "documento_numero",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "formato YYYY-MM-DD HH:MM:SS",
                        "name": "fecha_nacimiento",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Da de baja a un usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id del usuario que se desea dar de baja",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/usuarios/{id}": {
            "patch": {
                "description": "Actualiza parcialmente un usuario por su ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar atributos de usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del usuario a actualizar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de actualización en JSON",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ports.Patch"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Criptomoneda": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "simbolo": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "domain.Documento": {
            "type": "object",
            "properties": {
                "numero": {
                    "type": "string"
                },
                "tipo": {
                    "$ref": "#/definitions/domain.TipoDocumento"
                }
            }
        },
        "domain.TipoDocumento": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "DNI",
                "Cedula",
                "Pasaporte"
            ]
        },
        "domain.Usuario": {
            "type": "object",
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "documento": {
                    "$ref": "#/definitions/domain.Documento"
                },
                "email": {
                    "type": "string"
                },
                "fechaDeNacimiento": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monedasInteres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Criptomoneda"
                    }
                },
                "nombre": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "ports.Patch": {
            "type": "object",
            "properties": {
                "op": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "value": {}
            }
        }
    }
}